[#3857] filter & style fixes
